app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-storage==2.13.0
google-cloud-bigquery==3.13.0
